+++ R/tax_unique.R
@@ -29,16 +29,17 @@
 #' occurrence dataframe is of this type.
 #' @param resolution \code{character}. The taxonomic resolution at which to
 #' identify unique occurrences, either "species" (the default) or "genus".
-#' @param orig \code{logical}. Should the original dataframe be returned with
+#' @param append \code{logical}. Should the original dataframe be returned with
 #' the unique names appended as a new column?
 #'
 #' @return A \code{dataframe} of taxa, with each row corresponding to a unique
 #' "species" or "genus" in the dataset (depending on the chosen resolution).
 #' The dataframe will include the taxonomic information provided into the
 #' function, as well as a column providing the 'unique' names of each taxon. If
-#' \code{orig} is \code{TRUE}, the original dataframe (\code{occdf}) will be
+#' \code{append} is \code{TRUE}, the original dataframe (\code{occdf}) will be
 #' returned with these 'unique' names appended as a new column. Occurrences that
-#' do not represent unique taxa will have \code{NA} for their 'unique' names.
+#' are identified to a coarser taxonomic resolution which is already represented
+#' within the dataset will have \code{NA} for their 'unique' names.
 #'
 #' @details Palaeobiologists usually count unique taxa by retaining only
 #' unique occurrences identified to a given taxonomic resolution, however
@@ -105,6 +106,10 @@
 #' genera <- tax_unique(occdf = occdf, genus = "genus", family = "family",
 #' order = "order", class = "class", resolution = "genus")
 #'
+#' #Append unique names to the original occurrences
+#' genera_appended <- tax_unique(occdf = occdf, genus = "genus", family = "family",
+#' order = "order", class = "class", resolution = "genus", append = TRUE)
+#'
 #' #Create dataframe from lists
 #' occdf2 <- data.frame(species = c("rex", "aegyptiacus", NA), genus =
 #' c("Tyrannosaurus", "Spinosaurus", NA), family = c("Tyrannosauridae",
@@ -126,7 +131,7 @@
 #'
 tax_unique <- function(occdf = NULL, binomial = NULL, species = NULL,
                        genus = NULL, ..., name = NULL,
-                       resolution = "species", orig = FALSE) {
+                       resolution = "species", append = FALSE) {
   #Give errors for incorrect input
   if (is.null(occdf)) {
     stop("Must enter an `occdf` of occurrences or taxon names")
@@ -267,21 +272,18 @@ tax_unique <- function(occdf = NULL, binomial = NULL, species = NULL,
   }
 
   #Remove absolute repeats
-  if (orig) {
+  if (append) {
     occurrences_with_dupes <- occurrences
     # List the rows of the original dataframe that correspond to each unique row
     # Hack to include NAs as groups
     occurrences_with_dupes[is.na(occurrences_with_dupes)] <- "thisisanNAvalue"
     occurrences_with_dupes$rows <- seq_len(nrow(occurrences_with_dupes))
-    form <- as.formula(
-      paste(". ~",
-            paste(c(rev(higher_names), "genus", "genus_species"),
-                  collapse = " + "))
-      )
+    cols <- c(rev(higher_names), "genus")
+    if (resolution == "species") cols <- c(cols, "genus_species")
+    form <- as.formula(paste("rows ~", paste(cols, collapse = " + ")))
     occurrences <- aggregate(form, data = occurrences_with_dupes, FUN = c)
     # Switch hack groups back to true NAs
     occurrences[occurrences == "thisisanNAvalue"] <- NA
-
   } else {
     occurrences <- unique(occurrences)
   }
@@ -303,7 +305,7 @@ tax_unique <- function(occdf = NULL, binomial = NULL, species = NULL,
                               select = -c(genus_species)),
                        to_retain,
                        by = c(rev(higher_names), "genus"), all = T)
-    if (orig) {
+    if (append) {
       to_retain$rows <- ifelse(is.na(to_retain$rows.y),
                                to_retain$rows.x, to_retain$rows.y)
       to_retain$rows.x <- to_retain$rows.y <- NULL
@@ -311,14 +313,16 @@ tax_unique <- function(occdf = NULL, binomial = NULL, species = NULL,
     occurrences <- occurrences[is.na(occurrences$genus), ]
 
   } else if (resolution == "genus") {
-    #Remove genus_species column and remove genus repeats
-    if (!is.null(binomial) || !is.null(name)) {
-      occurrences <- subset(occurrences, select = -c(genus_species))
-    } else if (!is.null(species)) {
-      occurrences <- subset(occurrences, select = -c(genus_species, species))
-    }
+    if (!append) {
+      #Remove genus_species column and remove genus repeats
+      if (!is.null(binomial) || !is.null(name)) {
+        occurrences <- subset(occurrences, select = -c(genus_species))
+      } else if (!is.null(species)) {
+        occurrences <- subset(occurrences, select = -c(genus_species, species))
+      }
 
-    occurrences <- unique(occurrences)
+      occurrences <- unique(occurrences)
+    }
 
     #Retain genus identifications and remove from dataframe
     to_retain <- rbind(to_retain, occurrences[!is.na(occurrences$genus), ])
@@ -356,7 +360,7 @@ tax_unique <- function(occdf = NULL, binomial = NULL, species = NULL,
   }
 
   #Reorder
-  rows_col <- if (orig) "rows" else NULL
+  rows_col <- if (append) "rows" else NULL
   if (resolution == "species") {
     to_retain <- to_retain[, c(rev(higher_names), "genus", "genus_species",
                                "unique_name", rows_col)]
@@ -374,7 +378,7 @@ tax_unique <- function(occdf = NULL, binomial = NULL, species = NULL,
     to_retain <- to_retain[order(to_retain$genus_species), ]
   }
 
-  if (orig) {
+  if (append) {
     # Convert back to original dataframe
     occdf$unique_name <- NA
     for (i in seq_len(nrow(to_retain))) {